Refactoring code
Moving code mostly associated with discovery service.
Separated this change from the initial PR for a cleaner view of the work.

LEARNER-1135
