+++ ecommerce/enterprise/tests/test_entitlements.py
@@ -8,10 +8,8 @@ from slumber.exceptions import SlumberBaseException
 from testfixtures import LogCapture
 
 from ecommerce.core.tests import toggle_switch
-from ecommerce.core.tests.decorators import mock_course_catalog_api_client, mock_enterprise_api_client
 from ecommerce.coupons.tests.mixins import CouponMixin, CourseCatalogMockMixin
 from ecommerce.courses.tests.factories import CourseFactory
-from ecommerce.courses.tests.mixins import CourseCatalogServiceMockMixin
 from ecommerce.enterprise.entitlements import (
     get_course_entitlements_for_learner, get_course_vouchers_for_learner, get_entitlement_voucher,
     is_course_in_enterprise_catalog
@@ -28,8 +26,8 @@ StockRecord = get_model('partner', 'StockRecord')
 
 @ddt.ddt
 @httpretty.activate
-class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixin, CourseCatalogTestMixin,
-                        CourseCatalogMockMixin, CouponMixin, TestCase):
+class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogTestMixin, CourseCatalogMockMixin,
+                        CouponMixin, TestCase):
     def setUp(self):
         super(EntitlementsTests, self).setUp()
         self.learner = self.create_user(is_staff=True)
@@ -116,13 +114,13 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
         )
         return course_catalog_coupon
 
-    @mock_enterprise_api_client
     def test_get_entitlement_voucher_with_enterprise_feature_disabled(self):
         """
         Verify that method "get_entitlement_voucher" doesn't call the
         enterprise service API and returns no voucher if the enterprise
         feature is disabled.
         """
+        self.mock_access_token_response()
         self.mock_enterprise_learner_api()
         self.mock_enterprise_learner_entitlements_api()
         toggle_switch(settings.ENABLE_ENTERPRISE_ON_RUNTIME_SWITCH, False)
@@ -131,13 +129,12 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
         self._assert_num_requests(0)
         self.assertIsNone(entitlement_voucher)
 
-    @mock_enterprise_api_client
-    @mock_course_catalog_api_client
     def test_get_entitlement_voucher_with_enterprise_feature_enabled(self):
         """
         Verify that method "get_entitlement_voucher" returns a voucher if
         the enterprise feature is enabled.
         """
+        self.mock_access_token_response()
         coupon = self.create_coupon(catalog=self.catalog)
         expected_voucher = coupon.attr.coupon_vouchers.vouchers.first()
 
@@ -150,11 +147,9 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
         )
 
         entitlement_voucher = get_entitlement_voucher(self.request, self.course.products.first())
-        self._assert_num_requests(3)
+        self._assert_num_requests(4)
         self.assertEqual(expected_voucher, entitlement_voucher)
 
-    @mock_enterprise_api_client
-    @mock_course_catalog_api_client
     def test_get_entitlement_voucher_with_invalid_entitlement_id(self):
         """
         Verify that method "get_entitlement_voucher" logs exception if there
@@ -162,6 +157,7 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
         learner API response.
         """
         non_existing_coupon_id = 99
+        self.mock_access_token_response()
         self.mock_enterprise_learner_api(
             catalog_id=non_existing_coupon_id, entitlement_id=non_existing_coupon_id
         )
@@ -174,7 +170,7 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
         logger_name = 'ecommerce.enterprise.entitlements'
         with LogCapture(logger_name) as logger:
             entitlement_voucher = get_entitlement_voucher(self.request, self.course.products.first())
-            self._assert_num_requests(3)
+            self._assert_num_requests(4)
 
             logger.check(
                 (
@@ -185,8 +181,6 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
             )
             self.assertIsNone(entitlement_voucher)
 
-    @mock_enterprise_api_client
-    @mock_course_catalog_api_client
     def test_get_course_vouchers_for_learner_with_multiple_codes(self):
         """
         Verify that method "get_course_vouchers_for_learner" returns all valid
@@ -197,6 +191,7 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
         coupon = self._create_course_catalog_coupon(catalog_id, coupon_quantity)
         expected_vouchers = coupon.attr.coupon_vouchers.vouchers.all()
 
+        self.mock_access_token_response()
         self.mock_enterprise_learner_api(entitlement_id=coupon.id)
         self.mock_enterprise_learner_entitlements_api(entitlement_id=coupon.id)
         self.mock_course_discovery_api_for_catalog_contains(
@@ -209,45 +204,44 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
         # enterprise learner data and enterprise learner entitlements
         # from enterprise service and one for checking course run against
         # the enterprise catalog from the course catalog service.
-        self._assert_num_requests(3)
+        self._assert_num_requests(4)
         self.assertEqual(coupon_quantity, len(course_vouchers))
         self.assertListEqual(list(expected_vouchers), list(course_vouchers))
 
-    @mock_enterprise_api_client
     def test_get_course_vouchers_for_learner_with_exception(self):
         """
         Verify that method "get_course_vouchers_for_learner" returns empty
         response if there is an error while accessing the enterprise learner
         API.
         """
+        self.mock_access_token_response()
         self.mock_enterprise_learner_api_for_failure()
 
         vouchers = get_course_vouchers_for_learner(self.request.site, self.request.user, self.course.id)
         self.assertIsNone(vouchers)
 
-    @mock_enterprise_api_client
     def test_get_course_entitlements_for_learner_with_exception(self):
         """
         Verify that method "get_course_entitlements_for_learner" logs exception if
         there is an error while accessing the enterprise learner API.
         """
+        self.mock_access_token_response()
         self.mock_enterprise_learner_api_for_failure()
 
         self._assert_get_course_entitlements_for_learner_response(
             expected_entitlements=None,
             log_level='ERROR',
             log_message='Failed to retrieve enterprise info for the learner [%s]' % self.learner.username,
-            expected_request_count=1,
+            expected_request_count=2,
         )
 
-    @mock_enterprise_api_client
-    @mock_course_catalog_api_client
     def test_learner_entitlements_with_exception(self):
         """
         Verify that method "get_course_entitlements_for_learner" logs exception if
         there is an error while accessing the learner entitlements.
         """
         learner_id = 1
+        self.mock_access_token_response()
         self.mock_enterprise_learner_api(learner_id=learner_id)
         self.mock_learner_entitlements_api_failure(learner_id=learner_id)
         self.mock_course_discovery_api_for_catalog_contains(
@@ -258,17 +252,16 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
             expected_entitlements=None,
             log_level='ERROR',
             log_message='Failed to retrieve entitlements for enterprise learner [%s].' % learner_id,
-            expected_request_count=3,
+            expected_request_count=4,
         )
 
-    @mock_enterprise_api_client
-    @mock_course_catalog_api_client
     def test_learner_entitlements_invalid_response(self):
         """
         Verify that method "get_course_entitlements_for_learner" logs exception if
         there is an error while accessing the learner entitlements.
         """
         learner_id = 1
+        self.mock_access_token_response()
         self.mock_enterprise_learner_api(learner_id=learner_id)
         self.mock_learner_entitlements_api_failure(learner_id=learner_id, status=200)
         self.mock_course_discovery_api_for_catalog_contains(
@@ -280,31 +273,31 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
             log_level='ERROR',
             log_message='Invalid structure for enterprise learner entitlements API response for enterprise learner'
                         ' [%s].' % learner_id,
-            expected_request_count=3,
+            expected_request_count=4,
         )
 
-    @mock_enterprise_api_client
     def test_get_course_entitlements_for_learner_with_no_enterprise(self):
         """
         Verify that method "get_course_entitlements_for_learner" logs and returns
         empty list if the learner is not affiliated with any enterprise.
         """
+        self.mock_access_token_response()
         self.mock_enterprise_learner_api_for_learner_with_no_enterprise()
 
         self._assert_get_course_entitlements_for_learner_response(
             expected_entitlements=None,
             log_level='INFO',
             log_message='Learner with username [%s] in not affiliated with any enterprise' % self.learner.username,
-            expected_request_count=1,
+            expected_request_count=2,
         )
 
-    @mock_enterprise_api_client
     def test_get_course_entitlements_for_learner_with_invalid_response(self):
         """
         Verify that method "get_course_entitlements_for_learner" logs and returns
         empty list for entitlements if the enterprise learner API response has
         invalid/unexpected structure.
         """
+        self.mock_access_token_response()
         self.mock_enterprise_learner_api_for_learner_with_invalid_response()
 
         message = 'Invalid structure for enterprise learner API response for the learner [%s]' % self.learner.username
@@ -312,16 +305,16 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
             expected_entitlements=None,
             log_level='ERROR',
             log_message=message,
-            expected_request_count=1,
+            expected_request_count=2,
         )
 
-    @mock_enterprise_api_client
     def test_get_course_entitlements_for_learner_with_invalid_entitlements_key_in_response(self):
         """
         Verify that method "get_course_entitlements_for_learner" logs and returns
         empty list for entitlements if the enterprise learner API response has
         invalid/unexpected or missing key for enterprise customer entitlements.
         """
+        self.mock_access_token_response()
         self.mock_enterprise_learner_api_for_learner_with_invalid_entitlements_response()
 
         message = 'Invalid structure for enterprise learner API response for the learner [%s]' % self.learner.username
@@ -329,11 +322,9 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
             expected_entitlements=None,
             log_level='ERROR',
             log_message=message,
-            expected_request_count=1,
+            expected_request_count=2,
         )
 
-    @mock_enterprise_api_client
-    @mock_course_catalog_api_client
     def test_get_course_entitlements_for_learner_with_unavailable_course(self):
         """
         Verify that method "get_course_entitlements_for_learner" returns empty
@@ -341,6 +332,7 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
         in the related enterprise course catalog.
         """
         enterprise_catalog_id = 1
+        self.mock_access_token_response()
         self.mock_enterprise_learner_api(entitlement_id=enterprise_catalog_id)
         self.mock_enterprise_learner_entitlements_api(entitlement_id=enterprise_catalog_id)
         self.mock_course_discovery_api_for_catalog_contains(
@@ -356,16 +348,16 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
         # enterprise learner data from enterprise service and other one for
         # checking course run against the enterprise catalog query from the
         # course catalog service.
-        self._assert_num_requests(2)
+        self._assert_num_requests(3)
         self.assertIsNone(entitlements)
 
-    @mock_course_catalog_api_client
     def test_is_course_in_enterprise_catalog_for_available_course(self):
         """
         Verify that method "is_course_in_enterprise_catalog" returns True if
         the provided course is available in the enterprise course catalog.
         """
         enterprise_catalog_id = 1
+        self.mock_access_token_response()
         self.mock_course_discovery_api_for_catalog_contains(
             discovery_api_url=self.site_configuration.discovery_api_url, catalog_id=enterprise_catalog_id,
             course_run_ids=[self.course.id]
@@ -375,16 +367,16 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
 
         # Verify that there only one call for the course discovery API for
         # checking if course exists in course runs against the course catalog.
-        self._assert_num_requests(1)
+        self._assert_num_requests(2)
         self.assertTrue(is_course_available)
 
-    @mock_course_catalog_api_client
     def test_is_course_in_enterprise_catalog_for_unavailable_course(self):
         """
         Verify that method "is_course_in_enterprise_catalog" returns False if
         the provided course is not available in the enterprise course catalog.
         """
         enterprise_catalog_id = 1
+        self.mock_access_token_response()
         self.mock_course_discovery_api_for_catalog_contains(
             discovery_api_url=self.site_configuration.discovery_api_url, catalog_id=enterprise_catalog_id,
             course_run_ids=[self.course.id]
@@ -395,10 +387,9 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
 
         # Verify that there only one call for the course discovery API for
         # checking if course exists in course runs against the course catalog.
-        self._assert_num_requests(1)
+        self._assert_num_requests(2)
         self.assertFalse(is_course_available)
 
-    @mock_course_catalog_api_client
     @ddt.data(ConnectionError, SlumberBaseException, Timeout)
     def test_is_course_in_enterprise_catalog_for_error_in_get_course_catalogs(self, error):
         """
@@ -406,14 +397,12 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
         and logs error message if the method "get_course_catalogs" is unable
         to fetch catalog against the provided enterprise course catalog id.
         """
-        enterprise_catalog_id = 1
-        self.mock_catalog_api_failure(error, self.site_configuration.discovery_api_url, enterprise_catalog_id)
+        self.mock_access_token_response()
+        self.mock_catalog_api_failure(error, self.site_configuration.discovery_api_url, 1)
 
-        expected_number_of_requests = 1
         log_message = 'Unable to connect to Course Catalog service for catalog contains endpoint.'
-        self._assert_is_course_in_enterprise_catalog_for_failure(expected_number_of_requests, log_message)
+        self._assert_is_course_in_enterprise_catalog_for_failure(2, log_message)
 
-    @mock_course_catalog_api_client
     @ddt.data(ConnectionError, SlumberBaseException, Timeout)
     def test_is_course_in_enterprise_catalog_for_error_in_get_catalog_course_runs(self, error):
         """
@@ -422,12 +411,11 @@ class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogServiceMockMixi
         unable to validate the given course against the course runs for the
         provided enterprise catalog.
         """
-        enterprise_catalog_id = 1
+        self.mock_access_token_response()
         self.mock_get_catalog_contains_api_for_failure(
-            course_run_ids=[self.course.id], catalog_id=enterprise_catalog_id, error=error,
+            course_run_ids=[self.course.id], catalog_id=1, error=error,
             discovery_api_url=self.site_configuration.discovery_api_url
         )
 
-        expected_number_of_requests = 1
         log_message = 'Unable to connect to Course Catalog service for catalog contains endpoint.'
-        self._assert_is_course_in_enterprise_catalog_for_failure(expected_number_of_requests, log_message)
+        self._assert_is_course_in_enterprise_catalog_for_failure(2, log_message)
