+++ ecommerce/extensions/basket/tests/test_views.py
@@ -25,7 +25,6 @@ from waffle.testutils import override_flag
 from ecommerce.core.constants import ENROLLMENT_CODE_PRODUCT_CLASS_NAME, ENROLLMENT_CODE_SWITCH
 from ecommerce.core.exceptions import SiteConfigurationError
 from ecommerce.core.tests import toggle_switch
-from ecommerce.core.tests.decorators import mock_course_catalog_api_client
 from ecommerce.core.url_utils import get_lms_url
 from ecommerce.coupons.tests.mixins import CouponMixin, CourseCatalogMockMixin
 from ecommerce.courses.tests.factories import CourseFactory
@@ -491,12 +490,12 @@ class BasketSummaryViewTests(CourseCatalogTestMixin, CourseCatalogMockMixin, Lms
         (Benefit.FIXED, 50)
     )
     @ddt.unpack
-    @mock_course_catalog_api_client
     @override_settings(PAYMENT_PROCESSORS=['ecommerce.extensions.payment.tests.processors.DummyProcessor'])
     def test_response_success(self, benefit_type, benefit_value):
         """ Verify a successful response is returned. """
         seat = self.create_seat(self.course, 500)
         basket = self.create_basket_and_add_product(seat)
+        self.mock_access_token_response()
         self.create_and_apply_benefit_to_basket(basket, seat, benefit_type, benefit_value)
 
         self.assertEqual(basket.lines.count(), 1)
@@ -561,11 +560,11 @@ class BasketSummaryViewTests(CourseCatalogTestMixin, CourseCatalogMockMixin, Lms
         self.assertEqual(lines[0][1]['benefit_value'], '50%')
         self.assertEqual(lines[1][1]['benefit_value'], None)
 
-    @mock_course_catalog_api_client
     def test_cached_course(self):
         """ Verify that the course info is cached. """
         seat = self.create_seat(self.course, 50)
         basket = self.create_basket_and_add_product(seat)
+        self.mock_access_token_response()
         self.assertEqual(basket.lines.count(), 1)
         self.mock_dynamic_catalog_single_course_runs_api(
             self.course, discovery_api_url=self.site_configuration.discovery_api_url
@@ -590,11 +589,11 @@ class BasketSummaryViewTests(CourseCatalogTestMixin, CourseCatalogMockMixin, Lms
         'course': 'edX+DemoX',
         'short_description': None,
     })
-    @mock_course_catalog_api_client
     def test_empty_catalog_api_response(self, course_info):
         """ Check to see if we can handle empty response from the catalog api """
         seat = self.create_seat(self.course)
         self.create_basket_and_add_product(seat)
+        self.mock_access_token_response()
         self.mock_dynamic_catalog_single_course_runs_api(
             self.course, self.site_configuration.discovery_api_url, course_info
         )
@@ -687,10 +686,10 @@ class BasketSummaryViewTests(CourseCatalogTestMixin, CourseCatalogMockMixin, Lms
         ('2017-02-01T00:00:00', datetime.datetime(2017, 2, 1)),
     )
     @ddt.unpack
-    @mock_course_catalog_api_client
     @override_settings(PAYMENT_PROCESSORS=['ecommerce.extensions.payment.tests.processors.DummyProcessor'])
     def test_context_data_contains_course_dates(self, date_string, expected_result):
         seat = self.create_seat(self.course)
+        self.mock_access_token_response()
         self.create_basket_and_add_product(seat)
         self.mock_dynamic_catalog_single_course_runs_api(
             self.course,
