+++ ecommerce/extensions/offer/tests/test_models.py
@@ -12,10 +12,8 @@ from oscar.test import factories
 from requests.exceptions import ConnectionError, Timeout
 from slumber.exceptions import SlumberBaseException
 
-from ecommerce.core.tests.decorators import mock_course_catalog_api_client
 from ecommerce.core.utils import get_cache_key
 from ecommerce.coupons.tests.mixins import CouponMixin, CourseCatalogMockMixin
-from ecommerce.courses.tests.mixins import CourseCatalogServiceMockMixin
 from ecommerce.extensions.catalogue.tests.mixins import CourseCatalogTestMixin
 from ecommerce.tests.testcases import TestCase
 
@@ -26,7 +24,7 @@ Range = get_model('offer', 'Range')
 
 @ddt.ddt
 @httpretty.activate
-class RangeTests(CouponMixin, CourseCatalogServiceMockMixin, CourseCatalogTestMixin, CourseCatalogMockMixin, TestCase):
+class RangeTests(CouponMixin, CourseCatalogTestMixin, CourseCatalogMockMixin, TestCase):
     def setUp(self):
         super(RangeTests, self).setUp()
 
@@ -102,12 +100,12 @@ class RangeTests(CouponMixin, CourseCatalogServiceMockMixin, CourseCatalogTestMi
         with self.assertRaises(Exception):
             self.range.run_catalog_query(self.product)
 
-    @mock_course_catalog_api_client
     def test_run_catalog_query(self):
         """
         run_course_query() should return True for included course run ID's.
         """
         course, seat = self.create_course_and_seat()
+        self.mock_access_token_response()
         self.mock_dynamic_catalog_contains_api(
             query='key:*', course_run_ids=[course.id], discovery_api_url=self.site_configuration.discovery_api_url
         )
@@ -132,12 +130,12 @@ class RangeTests(CouponMixin, CourseCatalogServiceMockMixin, CourseCatalogTestMi
             cached_response = cache.get(cache_key)
             self.assertEqual(response, cached_response)
 
-    @mock_course_catalog_api_client
     def test_query_range_contains_product(self):
         """
         contains_product() should return the correct boolean if a product is in it's range.
         """
         course, seat = self.create_course_and_seat()
+        self.mock_access_token_response()
         self.mock_dynamic_catalog_contains_api(
             query='key:*', course_run_ids=[course.id], discovery_api_url=self.site_configuration.discovery_api_url
         )
@@ -150,7 +148,6 @@ class RangeTests(CouponMixin, CourseCatalogServiceMockMixin, CourseCatalogTestMi
         response = self.range.contains_product(seat)
         self.assertTrue(response)
 
-    @mock_course_catalog_api_client
     def test_course_catalog_query_range_contains_product(self):
         """
         Verify that the method "contains_product" returns True (boolean) if a
@@ -158,6 +155,7 @@ class RangeTests(CouponMixin, CourseCatalogServiceMockMixin, CourseCatalogTestMi
         """
         catalog_query = 'key:*'
         course, seat = self.create_course_and_seat()
+        self.mock_access_token_response()
         self.mock_dynamic_catalog_contains_api(
             query=catalog_query, course_run_ids=[course.id],
             discovery_api_url=self.site_configuration.discovery_api_url
@@ -180,9 +178,8 @@ class RangeTests(CouponMixin, CourseCatalogServiceMockMixin, CourseCatalogTestMi
         self.assertTrue(response)
         # Verify that there only one call for the course discovery API for
         # checking if course exists in course runs against the course catalog.
-        self._assert_num_requests(1)
+        self._assert_num_requests(2)
 
-    @mock_course_catalog_api_client
     @ddt.data(ConnectionError, SlumberBaseException, Timeout)
     def test_course_catalog_query_range_contains_product_for_failure(self, error):
         """
@@ -197,6 +194,7 @@ class RangeTests(CouponMixin, CourseCatalogServiceMockMixin, CourseCatalogTestMi
         self.range.course_catalog = course_catalog_id
         self.range.save()
 
+        self.mock_access_token_response()
         self.mock_catalog_api_failure(error, self.site_configuration.discovery_api_url, course_catalog_id)
         with self.assertRaises(Exception) as error:
             self.range.contains_product(seat)
@@ -205,9 +203,8 @@ class RangeTests(CouponMixin, CourseCatalogServiceMockMixin, CourseCatalogTestMi
         self.assertEqual(error.exception.message, expected_exception_message)
         # Verify that there only one call for the course discovery API for
         # checking if course exists in course runs against the course catalog.
-        self._assert_num_requests(1)
+        self._assert_num_requests(2)
 
-    @mock_course_catalog_api_client
     @ddt.data(
         'verified',
         'verified,professional',
@@ -226,6 +223,7 @@ class RangeTests(CouponMixin, CourseCatalogServiceMockMixin, CourseCatalogTestMi
         self.range.course_catalog = course_catalog
         self.range.save()
 
+        self.mock_access_token_response()
         self.mock_course_discovery_api_for_catalog_contains(
             discovery_api_url=self.site_configuration.discovery_api_url, catalog_id=course_catalog,
             course_run_ids=[course.id]
@@ -234,9 +232,8 @@ class RangeTests(CouponMixin, CourseCatalogServiceMockMixin, CourseCatalogTestMi
         self.assertTrue(is_product_in_range)
         # Verify that there only one call for the course discovery API for
         # checking if course exists in course runs against the course catalog.
-        self._assert_num_requests(1)
+        self._assert_num_requests(2)
 
-    @mock_course_catalog_api_client
     @ddt.data(
         ('verified', 'professional'),
         ('professional', 'verified'),
@@ -256,6 +253,7 @@ class RangeTests(CouponMixin, CourseCatalogServiceMockMixin, CourseCatalogTestMi
         self.range.course_catalog = course_catalog
         self.range.save()
 
+        self.mock_access_token_response()
         self.mock_course_discovery_api_for_catalog_contains(
             discovery_api_url=self.site_configuration.discovery_api_url, catalog_id=course_catalog,
             course_run_ids=[course.id]
@@ -266,7 +264,6 @@ class RangeTests(CouponMixin, CourseCatalogServiceMockMixin, CourseCatalogTestMi
         # doesn't have the provide course seat types.
         self._assert_num_requests(0)
 
-    @mock_course_catalog_api_client
     def test_query_range_all_products(self):
         """
         all_products() should return seats from the query.
@@ -275,6 +272,7 @@ class RangeTests(CouponMixin, CourseCatalogServiceMockMixin, CourseCatalogTestMi
         self.assertEqual(len(self.range.all_products()), 1)
         self.assertFalse(seat in self.range.all_products())
 
+        self.mock_access_token_response()
         self.mock_dynamic_catalog_course_runs_api(
             self.site_configuration.discovery_api_url, query='key:*', course_run=course
         )
