+++ ecommerce/tests/factories.py
@@ -1,6 +1,7 @@
 import factory
 from django.contrib.sites.models import Site
 from factory.fuzzy import FuzzyText     # pylint: disable=ungrouped-imports
+from faker import Faker
 from oscar.core.loading import get_model
 from oscar.test.factories import StockRecordFactory as OscarStockRecordFactory
 from oscar.test.factories import ProductFactory
@@ -36,9 +37,7 @@ class SiteConfigurationFactory(factory.DjangoModelFactory):
     send_refund_notifications = False
     enable_sdn_check = False
     enable_embargo_check = False
-    # TODO: Modify discovery_api_url to use Faker().url() and use
-    # mock_access_token_response instead of @mock_course_catalog_api_client decorator
-    discovery_api_url = factory.LazyAttribute(lambda obj: 'http://{domain}/api/v1/'.format(domain=obj.site.domain))
+    discovery_api_url = 'http://{}.fake/'.format(Faker().domain_name())
 
 
 class StockRecordFactory(OscarStockRecordFactory):
